AnswerService/requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.5.7
charset-normalizer==3.2.0
cloudinary==1.33.0
Django==4.1.9
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
idna==3.4
mysqlclient==2.2.0
PyJWT==1.7.1
pymongo==3.12.3
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.2.4
tzdata==2023.3